Self-normalized separated network module for GPD parameter prediction

A parameter-separated self-normalizing network as a torch::nn_module, designed for generalized Pareto distribution parameter prediction.

Usage

Separated_GPD_SNN(
  D_in,
  Hidden_vect_scale = c(64, 64, 64),
  Hidden_vect_shape = c(5, 3),
  p_drop = 0.01
)

Arguments

D_in

the input size (i.e. the number of features),

Hidden_vect_scale

a vector of integers whose length determines the number of layers in the sub-network for the scale parameter and entries the number of neurons in each corresponding successive layer,

Hidden_vect_shape

a vector of integers whose length determines the number of layers in the sub-network for the shape parameter and entries the number of neurons in each corresponding successive layer,

p_drop

probability parameter for the alpha-dropout before each hidden layer for regularization during training.

Value

The specified parameter-separated SNN MLP GPD network as a torch::nn_module.

Details

The constructor allows specifying:

D_in

the input size (i.e. the number of features),

Hidden_vect_scale

a vector of integers whose length determines the number of layers in the sub-network for the scale parameter and entries the number of neurons in each corresponding successive layer,

Hidden_vect_shape

a vector of integers whose length determines the number of layers in the sub-network for the shape parameter and entries the number of neurons in each corresponding successive layer,

p_drop

probability parameter for the alpha-dropout before each hidden layer for regularization during training.

References

Gunter Klambauer, Thomas Unterthiner, Andreas Mayr, Sepp Hochreiter. Self-Normalizing Neural Networks. Advances in Neural Information Processing Systems 30 (NIPS 2017), 2017.